requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scipy==1.11.0
pillow==10.0.0
google-cloud-vision==3.4.5
google-auth==2.23.4
protobuf==3.20.3
requests==2.31.0
